# Speed up DROP DATABASE